requirements.txt
torch
torchvision
numpy
pillow
opencv-python
streamlit
pyyaml
matplotlib
scikit-learn
tqdm
